code/pyGUIClient/requirements.txt
contourpy==1.0.7
cycler==0.11.0
fonttools==4.39.4
importlib-resources==5.12.0
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.3
packaging==23.1
Pillow==9.5.0
pip==23.0.1
pyparsing==3.0.9
python-dateutil==2.8.2
setuptools==67.8.0
six==1.16.0
wheel==0.38.4
zipp==3.15.0
